[BUGFIX] Always connect to dev machine
From the docs: xdebug.remote_connect_back: If enabled, the
xdebug.remote_host setting is ignored and Xdebug will try to connect to
the client that made the HTTP request.

But we always want the developers' machine to be contacted for PHP
debugging, even if the request comes from somewhere else - say e.g. a
mobile device, etc…